moved query list trade routes
