+++ x/stakeibc/client/cli/query.go
@@ -267,3 +267,28 @@ func CmdNextPacketSequence() *cobra.Command {
 
 	return cmd
 }
+
+func CmdListTradeRoutes() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "list-trade-routes",
+		Short: "list all trade routes",
+		RunE: func(cmd *cobra.Command, args []string) error {
+			clientCtx := client.GetClientContextFromCmd(cmd)
+
+			queryClient := types.NewQueryClient(clientCtx)
+
+			params := &types.QueryAllTradeRoutes{}
+			res, err := queryClient.AllTradeRoutes(context.Background(), params)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(res)
+		},
+	}
+
+	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
+	flags.AddQueryFlagsToCmd(cmd)
+
+	return cmd
+}
